Move zeff and dilution calculation to new formulas
