+++ cfspopcon/algorithms.yaml
@@ -420,7 +420,7 @@ calc_vertical_minor_radius_from_elongation_and_minor_radius:
   optionals: []
   returns:
   - vertical_minor_radius
-calc_zeff_and_dilution_from_impurities:
+calc_zeff_and_dilution_due_to_impurities:
   inputs:
   - average_electron_density
   - average_electron_temp
