+++ example_cases/SPARC_PRD/input.yaml
@@ -14,7 +14,7 @@ algorithms:
   - calc_q_star_from_plasma_current
   - calc_average_fuel_ion_mass
   - calc_average_ion_temp_from_temperature_ratio
-  - calc_zeff_and_dilution_from_impurities
+  - calc_zeff_and_dilution_due_to_impurities
   - calc_plasma_stored_energy
   - calc_power_balance_from_tau_e
   # Once we have P_in (=P_loss in steady-state), we want to split
